+++ doc/changelog.md
@@ -149,6 +149,7 @@ v0.6.6
 + (done, tested) Widget rendering is now doesn't depend on session, ommiting possibility of clutter display
 + (done, tested) Per subsite module & theme configuration 
 + (done, tested) Subsite module & theme privilege mechanism
++ (proposed) nordrassil priority index and table prefix
 + (proposed) namespace
 + (proposed) IP Deny manager.
 + (proposed) Navigation: move to anywhere.